Test for exclude/include error
I exclude mStudent.Cohort (which is id 26) and include Cohort id=809.
When I look at the object browser, you see 2 entities for
mStudent.Cohort although
I can not position on the excluded Cohort. But when we save to the
database,
the original cohort (id=26) gets saved not the new included cohort
(id=809).